# v-hormiga-app-main/src/documentation/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MIKNE API",
    "description": "API for MIKNE application",
    "contact": {
      "name": "Support Team",
      "email": "lazcano@icloud.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Cotizacion": {
      "post": {
        "tags": [
          "Cotizacion"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CotizacionCompleta"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CotizacionCompleta"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CotizacionCompleta"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CotizacionCompleta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Cotizacion"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CotizacionEncabezado"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CotizacionEncabezado"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CotizacionEncabezado"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Cotizacion/{id}": {
      "get": {
        "tags": [
          "Cotizacion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CotizacionCompleta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CotizacionCompleta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CotizacionCompleta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cotizacion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CotizacionCompleta"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CotizacionCompleta"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CotizacionCompleta"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CotizacionCompleta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Cotizacion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BahiaBasica": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "alimentacion": {
            "type": "boolean"
          },
          "riel": {
            "type": "boolean"
          },
          "estructura": {
            "type": "boolean"
          },
          "edited": {
            "type": "boolean"
          },
          "definiciones": {
            "$ref": "#/components/schemas/CotizacionBahiaDefiniciones"
          }
        },
        "additionalProperties": false
      },
...
      "ConceptoPrecio": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "number",
            "format": "double"
          },
          "precioUnitario": {
            "type": "number",
            "format": "double"
          },
          "importe": {
            "type": "number",
            "format": "double"
          },
          "tipo": {
            "type": "string",
            "nullable": true
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "observaciones": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CotizacionBahiaCompleta": {
        "type": "object",
        "properties": {
          "bahia": {
            "$ref": "#/components/schemas/BahiaBasica"
          }
        },
        "additionalProperties": false
      },
      "CotizacionBahiaDefiniciones": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idCotizacionBahia": {
            "type": "integer",
            "format": "int32"
          },
          "alimentacion": {
            "$ref": "#/components/schemas/CotizacionBahiaAlimentacion"
          },
          "riel": {
            "$ref": "#/components/schemas/CotizacionBahiaRiel"
          },
          "estructura": {
            "$ref": "#/components/schemas/CotizacionBahiaEstructura"
          },
          "cotizacionBahia": {
            "$ref": "#/components/schemas/CotizacionBahia"
          }
        },
        "additionalProperties": false
      },
...
      "CotizacionCompleta": {
        "type": "object",
        "properties": {
          "encabezado": {
            "$ref": "#/components/schemas/CotizacionEncabezado"
          },
          "productos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CotizacionProductoCompleto"
            },
            "nullable": true
          },
          "bahias": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CotizacionBahiaCompleta"
            },
            "nullable": true
          },
          "conceptos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConceptoPrecio"
            },
            "nullable": true
          },
          "resumen": {
            "$ref": "#/components/schemas/CotizacionResumen"
          },
          "metadata": {
            "$ref": "#/components/schemas/CotizacionMetadata"
          }
        },
        "additionalProperties": false
      },
      "CotizacionEncabezado": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tipoCotizacion": {
            "type": "string",
            "nullable": true
          },
          "tipoCuenta": {
            "type": "string",
            "nullable": true
          },
          "idiomaCotizacion": {
            "type": "string",
            "nullable": true
          },
          "cliente": {
            "type": "string",
            "nullable": true
          },
          "clienteNombre": {
            "type": "string",
            "nullable": true
          },
          "contacto": {
            "type": "string",
            "nullable": true
          },
          "dirFiscal": {
            "type": "string",
            "nullable": true
          },
          "dirEntrega": {
            "type": "string",
            "nullable": true
          },
          "referencia": {
            "type": "string",
            "nullable": true
          },
          "terminosEntrega": {
            "type": "string",
            "nullable": true
          },
          "folioPortal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "folioSap": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "vencimiento": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "moneda": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "impuestos": {
            "type": "number",
            "format": "double"
          },
          "descuentos": {
            "type": "number",
            "format": "double"
          },
          "observacionesGenerales": {
            "type": "string",
            "nullable": true
          },
          "condicionesComerciales": {
            "type": "string",
            "nullable": true
          },
          "vigencia": {
            "type": "integer",
            "format": "int32"
          },
          "creadoPor": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modificadoPor": {
            "type": "string",
            "nullable": true
          },
          "fechaModificacion": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CotizacionMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idCotizacion": {
            "type": "integer",
            "format": "int32"
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "fechaModificacion": {
            "type": "string",
            "format": "date-time"
          },
          "creadoPor": {
            "type": "string",
            "nullable": true
          },
          "modificadoPor": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "versionDocumento": {
            "type": "integer",
            "format": "int32"
          },
          "comentarios": {
            "type": "string",
            "nullable": true
          },
          "cotizacionEncabezado": {
            "$ref": "#/components/schemas/CotizacionEncabezado"
          }
        },
        "additionalProperties": false
      },
...
      "CotizacionProductoAdicionales": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idCotizacionProducto": {
            "type": "integer",
            "format": "int32"
          },
          "dispositivoTomaCarga": {
            "type": "boolean",
            "nullable": true
          },
          "carreteRetractil": {
            "type": "boolean",
            "nullable": true
          },
          "torreta": {
            "type": "boolean",
            "nullable": true
          },
          "especifiqueTorretaEspecial": {
            "type": "string",
            "nullable": true
          },
          "sirena": {
            "type": "boolean",
            "nullable": true
          },
          "tipoSirena": {
            "type": "string",
            "nullable": true
          },
          "especifiqueSirenaEspecial": {
            "type": "string",
            "nullable": true
          },
          "luminarias": {
            "type": "boolean",
            "nullable": true
          },
          "tipoLuminarias": {
            "type": "string",
            "nullable": true
          },
          "cantidadLuminarias": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "observaciones": {
            "type": "string",
            "nullable": true
          },
          "cotizacionProducto": {
            "$ref": "#/components/schemas/CotizacionProducto"
          }
        },
        "additionalProperties": false
      },
      "CotizacionProductoAlimentacion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idCotizacionProducto": {
            "type": "integer",
            "format": "int32"
          },
          "aliEleLarNave": {
            "type": "boolean",
            "nullable": true
          },
          "longitudSistema": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "acomElec": {
            "type": "string",
            "nullable": true
          },
          "localAcometida": {
            "type": "string",
            "nullable": true
          },
          "localAcometidaOtro": {
            "type": "string",
            "nullable": true
          },
          "amperaje": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "temperatura": {
            "type": "string",
            "nullable": true
          },
          "temperaturaOtro": {
            "type": "string",
            "nullable": true
          },
          "adecuadaAlimentar": {
            "type": "string",
            "nullable": true
          },
          "interruptorGeneral": {
            "type": "boolean",
            "nullable": true
          },
          "tipoAlim": {
            "type": "string",
            "nullable": true
          },
          "tipoAlimOtro": {
            "type": "string",
            "nullable": true
          },
          "areaTrabajo": {
            "type": "string",
            "nullable": true
          },
          "sistemaAlimentacionFestoon": {
            "type": "string",
            "nullable": true
          },
          "tipoSistemaAlimentacion": {
            "type": "string",
            "nullable": true
          },
          "observaciones": {
            "type": "string",
            "nullable": true
          },
          "cotizacionProducto": {
            "$ref": "#/components/schemas/CotizacionProducto"
          }
        },
        "additionalProperties": false
      },
      "CotizacionProductoCompleto": {
        "type": "object",
        "properties": {
          "producto": {
            "$ref": "#/components/schemas/ProductoBasico"
          }
        },
        "additionalProperties": false
      },
      "CotizacionProductoControl": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idCotizacionProducto": {
            "type": "integer",
            "format": "int32"
          },
          "medioControl": {
            "type": "string",
            "nullable": true
          },
          "fabricanteModelo": {
            "type": "string",
            "nullable": true
          },
          "observaciones": {
            "type": "string",
            "nullable": true
          },
          "tipoControl": {
            "type": "string",
            "nullable": true
          },
          "tipoCabina": {
            "type": "string",
            "nullable": true
          },
          "caracteristicasTecnicasCabina": {
            "type": "string",
            "nullable": true
          },
          "funcionesCabina": {
            "type": "string",
            "nullable": true
          },
          "funcionesAutomatizacion": {
            "type": "string",
            "nullable": true
          },
          "funcionesSemiautomatizacion": {
            "type": "string",
            "nullable": true
          },
          "marcaModeloPLC": {
            "type": "string",
            "nullable": true
          },
          "marcaModeloHMI": {
            "type": "string",
            "nullable": true
          },
          "cotizacionProducto": {
            "$ref": "#/components/schemas/CotizacionProducto"
          }
        },
        "additionalProperties": false
      },
...
      "CotizacionProductoDefiniciones": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idCotizacionProducto": {
            "type": "integer",
            "format": "int32"
          },
          "datosBasicos": {
            "$ref": "#/components/schemas/CotizacionProductoDatosBasicos"
          },
          "controles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CotizacionProductoControl"
            },
            "nullable": true
          },
          "adicionales": {
            "$ref": "#/components/schemas/CotizacionProductoAdicionales"
          },
          "izaje": {
            "$ref": "#/components/schemas/CotizacionProductoIzaje"
          },
          "alimentacion": {
            "$ref": "#/components/schemas/CotizacionProductoAlimentacion"
          },
          "tipoAlim": {
            "$ref": "#/components/schemas/CotizacionProductoTipoAlim"
          },
          "riel": {
            "$ref": "#/components/schemas/CotizacionProductoRiel"
          },
          "estructura": {
            "$ref": "#/components/schemas/CotizacionProductoEstructura"
          },
          "carros": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CarroIndividual"
            },
            "nullable": true
          },
          "puente": {
            "$ref": "#/components/schemas/CotizacionProductoPuente"
          },
          "flete": {
            "$ref": "#/components/schemas/CotizacionProductoFlete"
          },
          "montaje": {
            "$ref": "#/components/schemas/CotizacionProductoMontaje"
          },
          "informacionComplementaria": {
            "$ref": "#/components/schemas/CotizacionProductoInformacionComplementaria"
          },
          "formacionPrecios": {
            "$ref": "#/components/schemas/CotizacionProductoFormacionPrecios"
          },
          "cotizacionProducto": {
            "$ref": "#/components/schemas/CotizacionProducto"
          }
        },
        "additionalProperties": false
      },
...
      "CotizacionResumen": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idCotizacion": {
            "type": "integer",
            "format": "int32"
          },
          "subtotalProductos": {
            "type": "number",
            "format": "double"
          },
          "subtotalBahias": {
            "type": "number",
            "format": "double"
          },
          "subtotalConceptos": {
            "type": "number",
            "format": "double"
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "descuentoPorcentaje": {
            "type": "number",
            "format": "double"
          },
          "descuentoMonto": {
            "type": "number",
            "format": "double"
          },
          "subtotalConDescuento": {
            "type": "number",
            "format": "double"
          },
          "ivaPorcentaje": {
            "type": "number",
            "format": "double"
          },
          "ivaMonto": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "moneda": {
            "type": "string",
            "nullable": true
          },
          "tipoCambio": {
            "type": "number",
            "format": "double"
          },
          "observacionesFinancieras": {
            "type": "string",
            "nullable": true
          },
          "cotizacionEncabezado": {
            "$ref": "#/components/schemas/CotizacionEncabezado"
          }
        },
        "additionalProperties": false
      },
...
      "ProductoBasico": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "itemCode": {
            "type": "string",
            "nullable": true
          },
          "itemName": {
            "type": "string",
            "nullable": true
          },
          "qty": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "bahia": {
            "type": "string",
            "nullable": true
          },
          "definiciones": {
            "$ref": "#/components/schemas/CotizacionProductoDefiniciones"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}